Clean up code added in previous commits
