+++ dspace-server-webapp/src/test/data/dspaceFolder/config/spring/api/core-services-mock.xml
@@ -5,5 +5,5 @@
        default-lazy-init="true">
 
     <!-- Replace harvestCollectionService with a mocked service -->
-    <bean class="org.dspace.harvest.MockHarvestedCollectionServiceImpl" id="org.dspace.harvest.service.HarvestedCollectionService"/>
+    <bean class="org.dspace.harvest.MockHarvestedCollectionServiceImpl" id="org.dspace.harvest.service.HarvestedCollectionService" primary="true"/>
 </beans>
